requirements.txt
#Requirements
discord
validate_email_address
sqlalchemy==1.3.24
python-dotenv